refactor: content demo
