+++ src/assets/style/theme.scss
@@ -1,0 +1,3 @@
+.crx-class {
+  color: pink;
+}
