+++ src/pages/content/components/Demo/index.tsx
@@ -1,5 +1,5 @@
 import { createRoot } from "react-dom/client";
-import App from "@pages/contentView/app";
+import App from "@src/pages/content/components/Demo/app";
 
 const root = document.createElement("div");
 root.id = "chrome-extension-boilerplate-react-vite-content-view-root";
