+++ src/pages/content/index.ts
@@ -5,4 +5,4 @@ console.log("content loaded");
  * @description
  * Chrome extensions don't support modules in content scripts.
  */
-import("../contentView");
+import("./components/Demo");
