+++ src/pages/content/style.scss
@@ -1,3 +1,5 @@
+@import "@assets/style/theme.scss";
+
 .content-view {
   font-size: 30px;
 }
