+++ tsconfig.json
@@ -22,6 +22,5 @@
       "@pages/*": ["src/pages/*"]
     }
   },
-  "include": ["src",
-    "utils", "vite.config.ts"],
+  "include": ["src", "utils", "vite.config.ts"]
 }
